requirements.txt
requests
feedparser
Pillow
pocketsphinx
SpeechRecognition
pyaudio